omg-social-app: Refactor and optimize code in MessageBox and TextItems components
