+++ src/components/message-box/message-box-helper.js
@@ -1,0 +1,5 @@
+export const getRealMessage = (textObjInString) => {
+  const messageTextObj = JSON.parse(textObjInString);
+  const messageTextStr = messageTextObj.realTextMessage;
+  return messageTextStr;
+};
